requirements.txt
# ESP32 ProjectG MQTT 音訊系統 Python 依賴套件
paho-mqtt>=1.6.0
numpy>=1.21.0
sounddevice>=0.4.0
matplotlib>=3.5.0
pyaudio>=0.2.11
wave>=0.0.2
pyserial>=3.5
